+++ mex/editor/rules/main.py
@@ -30,6 +30,39 @@ def editor_value_switch(
     )
 
 
+def editor_edit_button(
+    field_name: str,
+    primary_source: EditorPrimarySource,
+    value: EditorValue,
+    index: int,
+) -> rx.Component:
+    """Return a button for toggling editing."""
+    return rx.icon_button(
+        rx.cond(
+            value.being_edited,
+            rx.icon(
+                "pencil-off",
+                height="1rem",
+                width="1rem",
+            ),
+            rx.icon(
+                "pencil",
+                height="1rem",
+                width="1rem",
+            ),
+        ),
+        variant="soft",
+        size="1",
+        on_click=[
+            RuleState.toggle_field_value_editing(field_name, index),
+            RuleState.resolve_identifiers,
+        ],
+        custom_attrs={
+            "data-testid": f"button-{field_name}-{primary_source.identifier}-{index}"
+        },
+    )
+
+
 def editor_static_value(
     field_name: str,
     primary_source: EditorPrimarySource,
@@ -45,6 +78,33 @@ def editor_static_value(
             value,
             index,
         ),
+    )
+
+
+def editor_additive_value(
+    field_name: str,
+    primary_source: EditorPrimarySource,
+    index: int,
+    value: EditorValue,
+) -> rx.Component:
+    """Render an additive value with buttons for editing and removal."""
+    return rx.hstack(
+        rx.cond(
+            value.being_edited,
+            additive_rule_input(
+                field_name,
+                primary_source.input_config,
+                index,
+                value,
+            ),
+            render_value(value),
+        ),
+        editor_edit_button(field_name, primary_source, value, index),
+        remove_additive_button(
+            field_name,
+            index,
+        ),
+        custom_attrs={"data-testid": f"additive-rule-{field_name}-{index}"},
     )
 
 
@@ -213,10 +273,6 @@ def additive_rule_input(
                 ),
             ),
         ),
-        remove_additive_button(
-            field_name,
-            index,
-        ),
     )
 
 
@@ -230,9 +286,9 @@ def editor_value_card(
     return rx.card(
         rx.cond(
             primary_source.input_config.allow_additive,
-            additive_rule_input(
+            editor_additive_value(
                 field_name,
-                primary_source.input_config,
+                primary_source,
                 index,
                 value,
             ),