+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- import artifical data helper and use in test
+- button for switching between rendered and input view for additive field values
+- import artificial data helper and use in test
 - page to create new rule-set in editor
 
 ### Changes